Adds bootstrap
This adds the bootstrap and jquery-rails gems to our application.  It
also configures the application to include them on all of our pages by
adding bootstrap to the application.js and application.scss files.  It
also updates the application.html.erb file to wrap all of our views in
some bootstraps div tags.